number and integer test conversions
